+++ Odotocodot.OneNote.Linq/Abstractions/IHasColor.cs
