+++ Odotocodot.OneNote.Linq/Abstractions/IHasIsInRecycleBin.cs
