+++ Odotocodot.OneNote.Linq/Abstractions/IHasPath.cs
