+++ Odotocodot.OneNote.Linq/Abstractions/INotebookOrSectionGroup.cs
