+++ Odotocodot.OneNote.Linq/IOneNoteItem.cs
