--- Odotocodot.OneNote.Linq/Interfaces/Abstractions/IWritableHasPath.cs
@@ -1,7 +1,0 @@
-namespace Odotocodot.OneNote.Linq.Abstractions
-{
-	internal interface IWritableHasPath : IHasPath
-	{
-		new string Path { set; }
-	}
-}
