+++ Odotocodot.OneNote.Linq/Internal/IWritableHasIsInRecycleBin.cs
@@ -1,4 +1,6 @@
-namespace Odotocodot.OneNote.Linq.Abstractions
+using Odotocodot.OneNote.Linq.Abstractions;
+
+namespace Odotocodot.OneNote.Linq.Internal
 {
 	internal interface IWritableHasIsInRecycleBin : IHasIsInRecycleBin
 	{
