+++ Odotocodot.OneNote.Linq/Internal/IWritableHasPath.cs
@@ -1,0 +1,9 @@
+using Odotocodot.OneNote.Linq.Abstractions;
+
+namespace Odotocodot.OneNote.Linq.Internal
+{
+    internal interface IWritableHasPath : IHasPath
+    {
+        new string Path { set; }
+    }
+}
